+++ RobotEdu-Client/src/main/java/robot/client/model/swipe/EduSwipeCardRecords.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public class EduSwipeCardRecords extends BaseModel {
 
-    private Long id;
+    private Integer id;
     private String cardNo;
     private String childName;
     private String upload;
@@ -26,11 +26,11 @@ public class EduSwipeCardRecords extends BaseModel {
         return Config.SERVER_URL + "";
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -69,7 +69,7 @@ public class EduSwipeCardRecords extends BaseModel {
 
     public static EduSwipeCardRecords getBean(Map<String, Object> item) {
         EduSwipeCardRecords bean = new EduSwipeCardRecords();
-        bean.setId((Long) item.get("id"));
+        bean.setId((Integer) item.get("id"));
         bean.setCardNo((String) item.get("card_no"));
         bean.setChildName((String) item.get("child_name"));
         bean.setUpload((String) item.get("upload"));
